[Wren:DS #35] Search for Ports 50 at a Time Instead of 1,000
I am not sure why it was previously necessary to skip over so many ports during a search. Since the RFC caps the port # range to 65,535, the old constant of 1,000 effectively limited the search space that the system would examine to no more than 64 ports (assuming the search started at <= 535).

Now, the search space is up to 1,310 ports.